Separate missed widgets into their own files Fixes #121
