+++ lib/widgets/auth/auth_page/images/crowd_league_logo.dart
