+++ lib/widgets/auth/auth_page/text/explanation_text.dart
