+++ lib/widgets/auth/email_auth_options_page/links/create_account_link.dart
@@ -1,0 +1,27 @@
+import 'package:crowdleague/actions/auth/update_email_auth_options_page.dart';
+import 'package:crowdleague/enums/auto_validate.dart';
+import 'package:crowdleague/enums/email_auth_mode.dart';
+import 'package:crowdleague/extensions/extensions.dart';
+import 'package:flutter/material.dart';
+
+class CreateAccountLink extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+        child: Text('CREATE ACCOUNT',
+            style: TextStyle(
+                fontSize: 15,
+                decoration: TextDecoration.underline,
+                color: Colors.blue)),
+        onTap: () {
+          Form.of(context).reset();
+          context.dispatch(UpdateEmailAuthOptionsPage(
+            mode: EmailAuthMode.signUp,
+            email: '',
+            password: '',
+            repeatPassword: '',
+            autovalidate: AutoValidate.disabled,
+          ));
+        });
+  }
+}
