+++ test/widget/auth/auth_page_test.dart
@@ -10,9 +10,9 @@ import 'package:crowdleague/widgets/auth/auth_page/buttons/auth_buttons.dart';
 import 'package:crowdleague/widgets/auth/auth_page/buttons/google_sign_in_fab.dart';
 import 'package:crowdleague/widgets/auth/auth_page/buttons/other_provider_fab.dart';
 import 'package:crowdleague/widgets/auth/auth_page/buttons/platform_sign_in_button.dart';
-import 'package:crowdleague/widgets/auth/auth_page/static_elements/crowd_league_logo.dart';
-import 'package:crowdleague/widgets/auth/auth_page/static_elements/explanation_text.dart';
-import 'package:crowdleague/widgets/auth/auth_page/static_elements/tag_line_text.dart';
+import 'package:crowdleague/widgets/auth/auth_page/images/crowd_league_logo.dart';
+import 'package:crowdleague/widgets/auth/auth_page/text/explanation_text.dart';
+import 'package:crowdleague/widgets/auth/auth_page/text/tag_line_text.dart';
 import 'package:crowdleague/widgets/shared/waiting_indicator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_auth_buttons/flutter_auth_buttons.dart';
